requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.9
fastapi==0.115.2
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.2
identify==2.6.1
idna==3.10
nodeenv==1.9.1
numpy==2.1.2
platformdirs==4.3.6
pre_commit==4.0.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.1
